[New V2]重写接收信息处理机制 有效提高反应速度 减少服务器压力
PS：与V1不兼容

+++ Wing.WeiXin.MP.SDK/Entities/SendAll/SendAllDelete.cs
@@ -14,14 +14,14 @@ namespace Wing.WeiXin.MP.SDK.Entities.SendAll
         /// <summary>
         /// 发送出去的消息ID
         /// </summary>
-        public int msgid { get; set; }
+        public long msgid { get; set; }
 
-        #region 根据发送出去的消息ID实例化 public SendAllDelete(int msgid)
+        #region 根据发送出去的消息ID实例化 public SendAllDelete(long msgid)
         /// <summary>
         /// 根据发送出去的消息ID实例化
         /// </summary>
         /// <param name="msgid">发送出去的消息ID</param>
-        public SendAllDelete(int msgid)
+        public SendAllDelete(long msgid)
         {
             this.msgid = msgid;
         }
